requirements.txt
pyaiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
environs==8.0.0
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
marshmallow==3.19.0
multidict==6.0.4
packaging==22.0
psycopg2==2.9.5
python-dotenv==0.21.0
pytube==12.1.2
pytz==2022.7
requests==2.28.1
soupsieve==2.3.2.post1
urllib3==1.26.13
yarl==1.8.2
